[refactor] enum 클래스 경로 수정
변경 사항 :
 - enum 클래스를 여러 도메인에서 사용하기 때문에 common/enumclass로 경로를 이동하였습니다.
