+++ src/main/java/com/final_10aeat/common/enumclass/ArticleCategory.java
@@ -1,4 +1,4 @@
-package com.final_10aeat.domain.repairArticle.entity;
+package com.final_10aeat.common.enumclass;
 
 public enum ArticleCategory {
     INSTALL, REPAIR, REPLACE
