+++ src/main/java/com/final_10aeat/common/enumclass/ManagePeriod.java
@@ -1,4 +1,4 @@
-package com.final_10aeat.domain.manageArticle.entity;
+package com.final_10aeat.common.enumclass;
 
 public enum ManagePeriod {
     YEAR, MONTH, WEEK, HALF
