+++ src/main/java/com/final_10aeat/common/enumclass/MemberRole.java
@@ -1,4 +1,4 @@
-package com.final_10aeat.domain.member.entity;
+package com.final_10aeat.common.enumclass;
 
 public enum MemberRole {
     ADMIN, OWNER, TENANT
