+++ src/main/java/com/final_10aeat/common/enumclass/Progress.java
@@ -1,4 +1,4 @@
-package com.final_10aeat.domain.repairArticle.entity;
+package com.final_10aeat.common.enumclass;
 
 public enum Progress {
     COMPLETE, INPROGRESS, PENDDING
